refactor: move PackageJson.ts inside `js-package-manager` directory
