+++ lib/cli/src/js-package-manager/PackageJson.ts
